Add options parameter, add `ignoreAutomaticGrants` option (#11)
* feat: add option to ignore automatic grants

* feat: merge `transform` into `userOptions`, add overloads to allow for any parameter combination

This allows for either providing just a transform function, just an options object with an (optional) transform function, or both as separate parameters. All parameters being optional!

* feat: remove 2-parameter overload

You can now either only pass a transform function, an object with a transform function and other options, or nothing at all.

As requested in a code review.